+++ plugin/src/main/kotlin/de/infolektuell/gradle/typst/extensions/ImageHandler.kt
@@ -1,0 +1,8 @@
+package de.infolektuell.gradle.typst.extensions
+
+import org.gradle.api.file.DirectoryProperty
+
+abstract class ImageHandler {
+    abstract val source: DirectoryProperty
+    abstract val converted: DirectoryProperty
+}